@@ -134,88 +141,77 @@
     "zxcvbn": "^4.4.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.3.4",
-    "@babel/plugin-proposal-class-properties": "^7.3.4",
-    "@babel/plugin-proposal-decorators": "^7.3.0",
-    "@babel/plugin-proposal-do-expressions": "^7.2.0",
-    "@babel/plugin-proposal-export-default-from": "^7.2.0",
-    "@babel/plugin-proposal-export-namespace-from": "^7.2.0",
-    "@babel/plugin-proposal-function-bind": "^7.2.0",
-    "@babel/plugin-proposal-function-sent": "^7.2.0",
-    "@babel/plugin-proposal-json-strings": "^7.2.0",
-    "@babel/plugin-proposal-logical-assignment-operators": "^7.2.0",
-    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.2.0",
-    "@babel/plugin-proposal-numeric-separator": "^7.2.0",
-    "@babel/plugin-proposal-optional-chaining": "^7.2.0",
-    "@babel/plugin-proposal-pipeline-operator": "^7.3.2",
-    "@babel/plugin-proposal-throw-expressions": "^7.2.0",
-    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "@babel/plugin-syntax-import-meta": "^7.2.0",
-    "@babel/plugin-transform-runtime": "^7.3.4",
-    "@babel/preset-env": "^7.3.4",
-    "@babel/register": "^7.0.0",
-    "@babel/runtime": "^7.3.4",
-    "babel-eslint": "^10.0.1",
-    "babel-loader": "^8.0.5",
+    "@babel/core": "^7.9.6",
+    "@babel/plugin-proposal-class-properties": "^7.8.3",
+    "@babel/plugin-proposal-decorators": "^7.8.3",
+    "@babel/plugin-proposal-do-expressions": "^7.8.3",
+    "@babel/plugin-proposal-export-default-from": "^7.8.3",
+    "@babel/plugin-proposal-export-namespace-from": "^7.8.3",
+    "@babel/plugin-proposal-function-bind": "^7.8.3",
+    "@babel/plugin-proposal-function-sent": "^7.8.3",
+    "@babel/plugin-proposal-json-strings": "^7.8.3",
+    "@babel/plugin-proposal-logical-assignment-operators": "^7.8.3",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.8.3",
+    "@babel/plugin-proposal-numeric-separator": "^7.8.3",
+    "@babel/plugin-proposal-optional-chaining": "^7.9.0",
+    "@babel/plugin-proposal-pipeline-operator": "^7.8.3",
+    "@babel/plugin-proposal-throw-expressions": "^7.8.3",
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
+    "@babel/plugin-syntax-import-meta": "^7.8.3",
+    "@babel/plugin-transform-runtime": "^7.9.6",
+    "@babel/preset-env": "^7.9.6",
+    "@babel/register": "^7.9.0",
+    "@babel/runtime": "^7.9.6",
+    "babel-loader": "^8.1.0",
     "babel-minify-webpack-plugin": "^0.3.1",
-    "babel-plugin-istanbul": "^5.1.1",
+    "babel-plugin-istanbul": "^5.2.0",
     "babili-webpack-plugin": "^0.1.2",
-    "cfonts": "^2.4.2",
+    "cfonts": "^2.8.2",
     "chai": "^4.2.0",
     "chalk": "^2.4.2",
-    "copy-webpack-plugin": "^5.0.0",
-    "cross-env": "^5.2.0",
-    "css-loader": "^2.1.0",
-    "del": "^4.0.0",
+    "copy-webpack-plugin": "^5.1.1",
+    "cross-env": "^5.2.1",
+    "css-loader": "^2.1.1",
+    "del": "^4.1.1",
     "devtron": "^1.4.0",
-    "electron": "^3.1.5",
+    "electron": "^3.1.13",
     "electron-builder": "21.2.0",
-    "electron-debug": "^2.1.0",
+    "electron-debug": "^2.2.0",
     "electron-toolkit": "^1.0.24",
-    "eslint": "^5.15.1",
-    "eslint-config-standard": "^12.0.0",
-    "eslint-friendly-formatter": "^4.0.1",
-    "eslint-loader": "^2.1.2",
-    "eslint-plugin-html": "^5.0.3",
-    "eslint-plugin-import": "^2.16.0",
-    "eslint-plugin-node": "^8.0.1",
-    "eslint-plugin-promise": "^4.0.1",
-    "eslint-plugin-standard": "^4.0.0",
-    "eslint-plugin-vue": "^5.2.2",
     "file-loader": "^3.0.1",
     "html-webpack-plugin": "^3.2.0",
     "inject-loader": "^4.0.1",
     "json-loader": "^0.5.7",
-    "karma": "^4.0.1",
+    "karma": "^4.4.1",
     "karma-chai": "^0.1.0",
     "karma-coverage": "^1.1.2",
-    "karma-electron": "^6.0.0",
+    "karma-electron": "^6.3.0",
     "karma-mocha": "^1.3.0",
     "karma-sourcemap-loader": "^0.3.7",
     "karma-spec-reporter": "^0.0.32",
     "karma-webpack": "^3.0.5",
     "mini-css-extract-plugin": "0.5.0",
-    "mocha": "^6.1.4",
+    "mocha": "^6.2.3",
     "multispinner": "^0.2.1",
     "node-loader": "^0.6.0",
-    "node-sass": "^4.12.0",
+    "node-sass": "^4.14.1",
     "require-dir": "^1.2.0",
-    "sass-loader": "^7.1.0",
+    "sass-loader": "^7.3.1",
     "spectron": "^5.0.0",
     "style-loader": "^0.23.1",
-    "ts-loader": "^6.0.0",
-    "typescript": "^3.4.5",
+    "ts-loader": "^6.2.2",
+    "typescript": "^3.8.3",
     "url-loader": "^1.1.2",
-    "vue-devtools": "^5.0.0-beta.1",
-    "vue-loader": "^15.7.0",
+    "vue-devtools": "^5.1.3",
+    "vue-loader": "^15.9.2",
     "vue-style-loader": "^4.1.2",
     "vue-svg-loader": "^0.12.0",
-    "vue-template-compiler": "^2.6.8",
-    "webpack": "^4.30.0",
-    "webpack-cli": "^3.2.3",
-    "webpack-dev-server": "^3.2.1",
-    "webpack-hot-middleware": "^2.24.3",
-    "webpack-merge": "^4.2.1",
+    "vue-template-compiler": "^2.6.11",
+    "webpack": "^4.43.0",
+    "webpack-cli": "^3.3.11",
+    "webpack-dev-server": "^3.10.3",
+    "webpack-hot-middleware": "^2.25.0",
+    "webpack-merge": "^4.2.2",
     "yaml-loader": "^0.5.0"
   }
 }
