@@ -20,13 +20,7 @@ let mainConfig = {
       {
         test: /\.(js)$/,
         enforce: 'pre',
-        exclude: /node_modules/,
-        use: {
-          loader: 'eslint-loader',
-          options: {
-            formatter: require('eslint-friendly-formatter')
-          }
-        }
+        exclude: /node_modules/
       },
       {
         test: /\.js$/,
