@@ -39,13 +39,7 @@ let rendererConfig = {
             {
                 test: /\.(js|vue)$/,
                 enforce: 'pre',
-                exclude: /node_modules/,
-                use: {
-                    loader: 'eslint-loader',
-                    options: {
-                        formatter: require('eslint-friendly-formatter')
-                    }
-                }
+                exclude: /node_modules/
             },
             {
                 test: /\.scss$/,
