@@ -1,22 +1,23 @@
 // Initialization routines for zcoind.
 
-import { Zcoind } from './zcoind';
+import { Zcoind, MnemonicSettings } from './zcoind';
 
-// For each component in src/lib/daemon/modules, register the exported function handleEvent() as an event handler for
-// the event with the name of the module, and also call the exported initialize() function.
-//
-// Each module may export a function handler:
-//
-//     async handleEvent(vuexStore: VuexStore, zcoind: Zcoind, eventData: any)
-//
-// and also may export an initializer:
-//
-//     async initialize(zcoind: Zcoind, store: VuexStore)
-//
-function zcoind(store: any): Zcoind {
-    // Note how this type differs from the one in zcoind.ts.
+/// Start up zcoind, connect to it, and return a Zcoind instance.
+async function zcoind(store: any, network: 'mainnet' | 'test' | 'regtest', zcoindLocation: string, zcoindDataDir: string,
+                      mnemonicSettings?: MnemonicSettings): Promise<Zcoind> {
+    // For each component in src/lib/daemon/modules, we register the exported function handleEvent() as an event handler for
+    // the event with the name of the module, and also call the exported initialize() function.
+    //
+    // Each module may export a function handler:
+    //
+    //     async handleEvent(vuexStore: VuexStore, zcoind: Zcoind, eventData: any)
+    //
+    // and also may export an initializer, which will be called after the block index is loaded:
+    //
+    //     async initialize(vuexStore: VuexStore, zcoind: Zcoind)
+    //
     const eventHandlers: {[topic: string]: (zcoind: Zcoind, eventData: any) => Promise<void>} = {};
-    const initializers: {[topic: string]: (vuexStore: any, zcoind: Zcoind) => Promise<void>} = {};
+    const initializers = [];
 
     const daemonComponents = require.context('./modules', true, /[^\/]\.ts$/);
     for (const fileName of daemonComponents.keys()) {
@@ -28,22 +29,18 @@ function zcoind(store: any): Zcoind {
         }
 
         if (component.initialize) {
-            initializers[topic] = component.initialize;
+            // Things won't work properly if component.initialize isn't an AsyncFunction.
+            if (component.initialize.constructor.name !== "AsyncFunction") {
+                throw `invalid initializer for ${topic}: initializer must be an async function`;
+            }
+
+            initializers.push(
+                async (zcoind) => await component.initialize(store, zcoind)
+            );
         }
     }
-
-    const zcoind = new Zcoind(eventHandlers);
-    zcoind.connectAndReact();
-
-    for (const topic of Object.keys(initializers)) {
-        // We want to do a specific check for this here because it's easy to get wrong, and if it's wrong, code will
-        // continue to work but just be slow.
-        if (initializers[topic].constructor.name !== "AsyncFunction") {
-            throw `invalid initializer for ${topic}: initializer must be an async function`;
-        }
-
-        initializers[topic](store, zcoind);
-    }
+    const zcoind = new Zcoind(network, zcoindLocation, zcoindDataDir, initializers, eventHandlers);
+    await zcoind.start(mnemonicSettings);
 
     return zcoind;
 }
