@@ -16,27 +16,18 @@
 </template>
 
 <script>
-import { mapGetters } from 'vuex'
 import GuideStepMixin from '@/mixins/GuideStepMixin'
 
 export default {
     name: 'IntroScreenWelcome',
+
     mixins: [
         GuideStepMixin
     ],
 
-    computed: {
-        ...mapGetters({
-            isInitialRun: 'App/isInitialRun'
-        })
-    },
-
     methods: {
-        isEnabled () {
-            return this.isInitialRun
-        },
         next() {
-            this.actions.next();
+            this.actions.goTo("location");
         }
     }
 }
