@@ -128,12 +128,7 @@ module.exports = {
         },
         testnet: undefined,
         type: 'full',
-        averageBlockTime: 0,
-        syncBlocksPerSecond: {
-            startBlockHeight: 0,
-            startTimestamp: 0,
-            currentTimestamp: 0
-        }
+        averageBlockTime: 0
     },
 
     Settings: { // daemon: isSubscription
